+++ src/nvim/globals.h
@@ -832,7 +832,7 @@ EXTERN char no_lines_msg[] INIT(= N_("--No lines in buffer--"));
 // When ":global" is used to number of substitutions and changed lines is
 // accumulated until it's finished.
 // Also used for ":spellrepall".
-EXTERN long sub_nsubs;       // total number of substitutions
+EXTERN int sub_nsubs;        // total number of substitutions
 EXTERN linenr_T sub_nlines;  // total number of lines changed
 
 // table to store parsed 'wildmode'
